DMT-54 refactor center-text logic into own class
Refactor all of the logic for center-text into centerText.js in order to encapsulate the logic and improve readability
